Separates test suites into dedicated files
